fix: modularizando sistema em blueprints
