+++ app/static/js/script-index.js
