+++ app/templates/base.html
@@ -15,7 +15,7 @@
 
 <header>
     <nav class="{% block class_navbar %}navbar{% endblock %}">
-        <p class="brand">Feira de Jaguaribe</p>
+        <a class="brand" href="{{ url_for('main.pagina_inicial' ) }}">Feira de Jaguaribe</a>
 
         <button class="botao">
             <span class="barra"></span>
@@ -24,17 +24,17 @@
         </button>
         <div class="navbar-links">
             <ul>
-                <li><a href="{{ url_for('pagina_inicial') }} ">Inicio</a></li>
-                <li><a href="{{ url_for('pagina_produtos') }}">Produtos</a></li>
+                <li><a href="{{ url_for('main.pagina_inicial') }} ">Inicio</a></li>
+                <li><a href="{{ url_for('main.pagina_produtos') }}">Produtos</a></li>
                 
                 {% if current_user.is_authenticated %}
-                    <li><a href="{{ url_for('pagina_perfil') }}" div class="button-perfil">
+                    <li><a href="{{ url_for('main.pagina_perfil') }}" div class="button-perfil">
                             <div class="icon-login">
                                 <i class="fa-regular fa-circle-user"></i>
                             </div>
                     </a></li>
                 {% else %}
-                    <li><a href="{{ url_for('pagina_login') }}" div class="button-login">
+                    <li><a href="{{ url_for('main.pagina_login') }}" div class="button-login">
                         <div class="icon-login">
                             <i class="fa-regular fa-circle-user"></i>
                         </div>
@@ -54,7 +54,6 @@
     <!-- Footer da Página -->
     <div class="footer">    
             <p class="descricao">Melhor plataforma para vendedores</p>
-
             <div class="direitos">
                 <p>&copy; 2025 Feira de jaguaribe | Todos os direitos reservados.</p>
             </div>
