+++ app/templates/cadastro_produtos.html
@@ -18,7 +18,7 @@
     <h1>Cadastro de Produto</h1>
     <div class="container">
 
-        <form action="{{url_for('cadastro_produtos')}}" method="post" enctype="multipart/form-data">
+        <form action="{{url_for('main.cadastro_produtos')}}" method="post" enctype="multipart/form-data">
 
             <div class="campo">
                 <label for="nome">Nome do Produto</label>
