+++ app/templates/index.html
@@ -21,7 +21,7 @@
     </div>
 
     <div class="botoes">
-        <a href="{{ url_for('pagina_produtos' ) }}" id="btn_ver_produtos">Ver Produtos</a>
+        <a href="{{ url_for('main.pagina_produtos' ) }}" id="btn_ver_produtos">Ver Produtos</a>
         <a href="#"id="btn_historia">Conhecer História</a>
     </div>
 </section>
@@ -63,7 +63,7 @@
         </div>
     </div>
 
-    <div class="direitos">
+    <div class="direitos_de_privacidade">
         <p>&copy 2025 Feira de Jaguaribe. Todos os direitos reservados.</p>
         <div id="termos">
             <p><a href="#">Privacidade</a></p>
