+++ app/templates/login.html
@@ -26,5 +26,5 @@
       </form>
     </div>
 
-    <p>Não tem uma conta? <a href="{{ url_for('cadastro') }}">Cadastre-se aqui</a></p>
+    <p>Não tem uma conta? <a href="{{ url_for('main.cadastro') }}">Cadastre-se aqui</a></p>
 {% endblock %}
