+++ data/usuarios.csv
@@ -1,4 +1,3 @@
 username;nome;data;whatsapp;senha
-software>hardware;software da silva junior;0001-01-01;None;pbkdf2:sha256:1000000$ki5HBt1GP1mCAvI0$a3700a2f0d81705b8e35b2442086617abe562a6c0506c9c22595f4f9b29f741c
+software>hardware;petronio;0001-01-01;None;pbkdf2:sha256:1000000$ki5HBt1GP1mCAvI0$a3700a2f0d81705b8e35b2442086617abe562a6c0506c9c22595f4f9b29f741c
 mengao;diogo presida;1981-12-13;5583993214976;pbkdf2:sha256:1000000$1K2SpC88pNzbphK0$8dc32ff0d2befc6cc377e1d2c75e963fe791cc6c07301b89f81a2a30befffd4e
-ryan;Ryan;2007-11-04;5581993724365;pbkdf2:sha256:1000000$aaafJtQyrQ9NbqAU$c8377b438f0ba88a47f799c828b3ed457791a3fa17b2b23057de8b5f75cad158
